refactoring

+++ web/app.py
@@ -2,76 +2,36 @@ from flask import Flask
 from flask import render_template
 from flask import request
 import re
-import random
-from paho.mqtt import client as mqtt_client
 import pandas as pd
 import os
-from sendgrid import SendGridAPIClient
-from sendgrid.helpers.mail import Mail
 
-broker = '192.168.148.238'
-port = 10000
+from constants.constants import SENDGRID_API_KEY, BROKER, PORT
+from services.send_email import send_message
+from services.mqtt import connect_mqtt
+
 topic = "test/topic"
 # generate client ID with pub prefix randomly
-client_id = f'python-mqtt-{random.randint(0, 100)}'
-username = ''
-password = ''
-user_email = 'camgcamg11@gmail.com'
 
-message = Mail(
-    from_email='johan.tanta@utec.edu.pe',
-    to_emails='cesar.madera@utec.edu.pe',
-    subject='Test Proyecto IOT',
-    html_content='<strong> Te están robando la hato, aiuda </strong>')
-from dotenv import load_dotenv
 
-load_dotenv()
+app = Flask(__name__)
+mqtt_client = None
 
-def send_message():
-    try:
-        sg = SendGridAPIClient(os.environ.get('SENDGRID_API_KEY'))
-        print(os.environ.get('SENDGRID_API_KEY'))
-        response = sg.send(message)
-        print(response.status_code)
-        print(response.body)
-        print(response.headers)
-    except Exception as e:
-        print(e)
-
-def connect_mqtt() -> mqtt_client:
-    def on_connect(client, userdata, flags, rc):
-        if rc == 0:
-            #print("Connected to MQTT Broker!")
-            next
-        else:
-            print("Failed to connect, return code %d\n", rc)
-
-    client = mqtt_client.Client(client_id)
-    #client.username_pw_set(username, password)
-    client.on_connect = on_connect
-    client.connect(broker, port)
-    return client
-
-def subscribe(client: mqtt_client):
+def subscribe(client, topic):
     def on_message(client, userdata, msg):
         message_decoded = msg.payload.decode()
         print(message_decoded)
         splitted_message = message_decoded.split()
+        
         if splitted_message[0] == "connect":
             with open('./data.csv','a') as f:
                 f.write(f'{splitted_message[1]}\n')
         elif splitted_message[0] == "alarma":
-            send_message()
+            send_message(SENDGRID_API_KEY)
         #print(f"Received `{msg.payload.decode()}` from `{msg.topic}` topic")
-
 
     client.subscribe(topic)
     client.on_message = on_message
-
-client = connect_mqtt()
-subscribe(client)
-client.loop_start()
-app = Flask(__name__)
+    
 
 def get_devices():
     df = pd.read_csv('./data.csv')
@@ -106,7 +66,9 @@ def validate_strings(tiempo,distancia,topic,email):
     return True 
 
 def create_file():
-    os.remove("./data.csv")
+    if os.path.exists('./data.csv'):
+        os.remove("./data.csv")
+    
     with open("./data.csv", "w") as f:
         f.write('ID\n')
         f.close()
@@ -125,7 +87,10 @@ def default():
 def index(name=None):
     return render_template('index.html', name=name)
 
-
+@app.route('/send_email')
+def send():
+    send_message(SENDGRID_API_KEY)
+    return "<p>Email enviado</p>"
 
 @app.route('/set_parameters')
 def set_parameters():
@@ -153,5 +118,9 @@ def set_parameters():
     return "<p> Not valid parameters </p>"
 
 if __name__ == '__main__':
-    create_file() 
+    mqtt_client = connect_mqtt(BROKER, PORT)
+    subscribe(mqtt_client, topic)
+    mqtt_client.loop_start()
+    
+    create_file()
     app.run(host="0.0.0.0", port=8080, threaded=True)
